requirements.txt
openai>=0.27.0
PyYAML>=6.0
pydantic>=1.10.0
jsonschema>=4.17.3
tiktoken
